refactor: 페이지 컴포넌트를 pages 폴더로 이동
- components/pages/ 폴더 생성
- 페이지 컴포넌트 5개 이동 (new-project-page, profile-page, settings-page, project-list-page, team-management-page)
- app 폴더 내 모든 import 경로 업데이트 (@/components/pages/로 변경)
