+++ app/(dashboard)/settings/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from "react"
 import { useRouter } from "next/navigation"
-import { SettingsPage } from "@/components/settings-page"
+import { SettingsPage } from "@/components/pages/settings-page"
 import { getCurrentProjectId } from "@/lib/storage"
 
 export default function Settings() {
